requirements.txt
# Core Requirements
ultralytics>=8.0.0  # YOLO framework
torch>=2.0.0  # PyTorch base (with CUDA if available)

# Video Processing
opencv-python>=4.7.0  # Video I/O and image processing

# Data Management
roboflow>=1.1.0  # Dataset download and management
pyyaml>=6.0  # Configuration file handling

# Visualization
matplotlib>=3.7.0  # Metrics plotting
seaborn>=0.12.2  # Enhanced visualizations

# Utilities
numpy>=1.24.0  # Array processing
tqdm>=4.65.0  # Progress bars
loguru>=0.7.0  # Enhanced logging (optional)

# Optional GPU Support
# torchvision>=0.15.0  # Uncomment if using custom transforms
# onnx>=1.13.0  # Uncomment for ONNX export